Clean test output, remove build warnings and unused code
