+++ cqc/rustLib/tests/run_test.sh
@@ -1,12 +1,16 @@
 #!/bin/bash
 
+set -eu
+
 # Start the Virtual and CQC Nodes
-../../../run/startVNodes.sh
-../../../run/startCQCNodes.sh
+echo -e "\e[1;32m[$(date +%H:%M:%S)] Start Virtual Nodes\e[0m"
+nohup $NETSIM/run/startVNodes.sh > /dev/null 2>&1 && sleep 5
+
+echo -e "\e[1;32m[$(date +%H:%M:%S)] Start CQC Nodes\e[0m"
+nohup $NETSIM/run/startCQCNodes.sh > /dev/null 2>&1 && sleep 5
 
 # start the test
-cargo test test
-cargo test test_gates
-cargo test test_send
-cargo test test_recv
+echo -e "\e[1;32m[$(date +%H:%M:%S)] Run Tests\e[0m"
+cargo test -- --nocapture
 
+echo -e "\e[1;32m[$(date +%H:%M:%S)] Testing Complete\e[0m"
