+++ cqc/rustLib/tests/test_init.rs
